pyproject.toml
[project]
name = "gh-pydantic-ai"
version = "0.3.0"
description = "GitHub Copilot Provider for Pydantic AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX :: Linux",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
]

dependencies = [
    "click>=8.2.1",
    "httpx>=0.28.1",
    "pydantic-ai-slim[openai,cli]>=1.2.0",
    "pyperclip>=1.9.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.1",
    "ty>=0.0.1a23",
]

[project.scripts]
gh-clai = "gh_pydantic_ai.cli:cli"
gh-clai-internal = "gh_pydantic_ai.cli:internal_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "ALL",  # select all checks by default
]
ignore = [
    # single rule exclude
    "TRY003",  # allow long exc message
    "COM812",  # will be handled by formatting
    "TC006",   # ignore quotes in typing.cast calls

    # full rule exclude
    "D",   # ignore docstrings
    "FA",  # ignore future annotations
    "ANN", # ignore annotations
    "FIX", # ignore FIXMEs
    "TD",  # ignore all TODOs
    "ARG", # ignore unused arguments
    "FBT", # ignore flake8-boolean-trap
    "EM",  # ignore error message formatting
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
]
per-file-ignores = { }
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
max-complexity = 10
